Renamed to cls_def
