+++ conf/script/src/build_system/cmd/clean_build_dir/cli.py
@@ -1,7 +1,7 @@
 import argparse
 
 import build_system.cmd.hierarchy.find_root_dir
-import utils.error.error
+import utils.error.cls_def
 import utils.error.status
 from build_system import cmd
 
@@ -15,7 +15,7 @@ def clean_build_dir():
     try:
         try:
             root_dir = cmd.hierarchy.find_root_dir.find_root_dir()
-        except utils.error.error.RootDirNotFoundError as raised_exception:
+        except utils.error.cls_def.RootDirNotFoundError as raised_exception:
             arg_parser.exit(utils.error.status.ErrorStatus.ROOT_DIR_NOT_FOUND, str(raised_exception))
 
         try:
