+++ conf/script/src/build_system/cmd/hierarchy/find_root_dir/cli.py
@@ -1,6 +1,6 @@
 import argparse
 
-import utils.error.error
+import utils.error.cls_def
 import utils.error.status
 from build_system import cmd
 from build_system.cmd.hierarchy.find_root_dir.find_root_dir import VCS_DIR_NAME
@@ -14,7 +14,7 @@ def find_root_dir():
         project_root = cmd.hierarchy.find_root_dir.find_root_dir()
         print(project_root, end=str())
 
-    except utils.error.error.RootDirNotFoundError as raised_exception:
+    except utils.error.cls_def.RootDirNotFoundError as raised_exception:
         arg_parser.exit(utils.error.status.ErrorStatus.ROOT_DIR_NOT_FOUND, str(raised_exception))
 
     except OSError as raised_exception:
