+++ conf/script/src/utils/error/synthesis.py
@@ -1,17 +1,17 @@
 from typing import Final, Type
 
-import utils.error.error
+import utils.error.cls_def
 import utils.error.status
 import utils.error.manage
 
 ALL_ERRORS: Final[list[Type[utils.cli.managed_error.ManagedError]]] = [
-    utils.error.error.SuccessWarning,
-    utils.error.error.UnsupportedError,
-    utils.error.error.ArgParserError,
-    utils.error.error.UnknownParsedArgError,
-    utils.error.error.EmptyParsedArgError,
-    utils.error.error.RootDirNotFoundError,
-    utils.error.error.BuildDirNotFoundError,
+    utils.error.cls_def.SuccessWarning,
+    utils.error.cls_def.UnsupportedError,
+    utils.error.cls_def.ArgParserError,
+    utils.error.cls_def.UnknownParsedArgError,
+    utils.error.cls_def.EmptyParsedArgError,
+    utils.error.cls_def.RootDirNotFoundError,
+    utils.error.cls_def.BuildDirNotFoundError,
 ]
 
 ALL_ERRORS_BY_STATUS: Final[dict[utils.error.status.ErrorStatus, Type[utils.cli.managed_error.ManagedError]]] = {status: ALL_ERRORS[status] for status in
